tagging/remote_functions/ultimate_source/requirements.txt
requests
google-auth-oauthlib